+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * The `api::paged` type now has an `.iter()` method which may be used to
     iterate over paginated results using lazily fetched API results. This can
     be used to reduce memory usage for large result sets.
+  * Added a `confidential` parameter for `api::projects::issues::EditIssue`
 
 # v0.1312.0